requirements.txt
paho-mqtt
flask
gunicorn